Moves Contentful resources to a sub-namespace
