+++ src/Management/Resource/Entry.php
@@ -7,9 +7,15 @@
  * @license   MIT
  */
 
-namespace Contentful\Management;
+namespace Contentful\Management\Resource;
 
 use Contentful\DateHelper;
+use Contentful\Management\Behavior\Archivable;
+use Contentful\Management\Behavior\Creatable;
+use Contentful\Management\Behavior\Deletable;
+use Contentful\Management\Behavior\Publishable;
+use Contentful\Management\Behavior\Updatable;
+use Contentful\Management\SystemProperties;
 
 /**
  * Entry class.
