+++ src/Management/Resource/PublishedContentType.php
@@ -7,7 +7,9 @@
  * @license   MIT
  */
 
-namespace Contentful\Management;
+namespace Contentful\Management\Resource;
+
+use Contentful\Management\PublishedSystemProperties;
 
 /**
  * PublishedContentType class.
