+++ src/Management/Resource/WebhookHealth.php
@@ -7,7 +7,9 @@
  * @license   MIT
  */
 
-namespace Contentful\Management;
+namespace Contentful\Management\Resource;
+
+use Contentful\Management\SystemProperties;
 
 /**
  * WebhookHealth class.
