+++ src/Management/ResourceBuilder.php
@@ -12,6 +12,19 @@ namespace Contentful\Management;
 use Contentful\File;
 use Contentful\Link;
 use Contentful\Management\Field\Validation;
+use Contentful\Management\Resource\Asset;
+use Contentful\Management\Resource\ContentType;
+use Contentful\Management\Resource\Entry;
+use Contentful\Management\Resource\EntrySnapshot;
+use Contentful\Management\Resource\Locale;
+use Contentful\Management\Resource\PublishedContentType;
+use Contentful\Management\Resource\Space;
+use Contentful\Management\Resource\Upload;
+use Contentful\Management\Resource\User;
+use Contentful\Management\Resource\Webhook;
+use Contentful\Management\Resource\WebhookCall;
+use Contentful\Management\Resource\WebhookCallDetails;
+use Contentful\Management\Resource\WebhookHealth;
 use Contentful\ResourceArray;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
