+++ tests/E2E/EntryTest.php
@@ -10,8 +10,8 @@
 namespace Contentful\Tests\E2E;
 
 use Contentful\Link;
-use Contentful\Management\Entry;
 use Contentful\Management\Query;
+use Contentful\Management\Resource\Entry;
 use Contentful\Tests\End2EndTestCase;
 
 class EntryTest extends End2EndTestCase
