+++ tests/E2E/SpaceTest.php
@@ -12,7 +12,7 @@ namespace Contentful\Tests\E2E;
 use Contentful\Exception\NotFoundException;
 use Contentful\Link;
 use Contentful\Management\Query;
-use Contentful\Management\Space;
+use Contentful\Management\Resource\Space;
 use Contentful\ResourceArray;
 use Contentful\Tests\End2EndTestCase;
