+++ tests/E2E/WebhookTest.php
@@ -10,10 +10,10 @@
 namespace Contentful\Tests\E2E;
 
 use Contentful\Link;
-use Contentful\Management\Entry;
 use Contentful\Management\Query;
-use Contentful\Management\Webhook;
-use Contentful\Management\WebhookCall;
+use Contentful\Management\Resource\Entry;
+use Contentful\Management\Resource\Webhook;
+use Contentful\Management\Resource\WebhookCall;
 use Contentful\Tests\End2EndTestCase;
 
 class WebhookTest extends End2EndTestCase
